src/test/java/com/example/project/Controllers/ReceiptControllerTest.java
package com.example.project.Controllers;

import com.example.project.controller.ReceiptController;
import com.example.project.model.Receipt;
import com.example.project.service.ReceiptService;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Collections;
import java.util.Optional;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(MockitoExtension.class)
public class ReceiptControllerTest {

    @Mock
    private ReceiptService receiptService;

    @InjectMocks
    private ReceiptController receiptController;

    private MockMvc mockMvc;
    private final ObjectMapper objectMapper = new ObjectMapper();

    @BeforeEach
    public void setup() {
        mockMvc = MockMvcBuilders.standaloneSetup(receiptController).build();
    }

    @Test
    public void testGetReceipts() throws Exception {
        when(receiptService.getAllReceipts()).thenReturn(Collections.emptyList());

        mockMvc.perform(get("/receipts")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
    }

    @Test
    public void testAddReceipt() throws Exception {
        Receipt mockReceipt = new Receipt(); // Consider setting necessary properties if needed.

        mockMvc.perform(post("/receipts")
                        .content(objectMapper.writeValueAsString(mockReceipt))
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isCreated());
    }

    @Test
    public void testGetReceiptById() throws Exception {
        when(receiptService.getReceiptById(1L)).thenReturn(Optional.of(new Receipt()));

        mockMvc.perform(get("/receipts/1")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
    }

    @Test
    public void testDeleteReceiptById() throws Exception {
        mockMvc.perform(delete("/receipts/1")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
    }
}
